refactor: 程序调通
