+++ src/Extensions/Components/BootstrapBlazor.DockView/wwwroot/js/dockview-fix.js
@@ -21,3 +21,5 @@ const removeInvisibleBranch = branch => {
         })
     }
 }
+
+export { fixObject };
