+++ src/Extensions/Components/BootstrapBlazor.DockView/wwwroot/js/dockview-panel.js
@@ -61,5 +61,23 @@ const findPanelFunc = v => p => findPanel(p, v);
 
 const findPanel = (p, v) => (p.params.key && p.params.key === v.params.key) || p.id === v.id || p.title === v.title;
 
+const savePanel = (dockview, panel) => {
+    const { panels, options } = dockview.params;
+    panels.push(panel)
+    if (options.enableLocalStorage) {
+        localStorage.setItem(`${options.localStorageKey}-panels`, JSON.stringify(panels))
+    }
+}
+
+const deletePanel = (dockview, panel) => {
+    const { panels, options } = dockview.params;
+    let index = panels.indexOf(panel);
+    if (index > -1) {
+        panels.splice(index, 1);
+    }
+    if (options.enableLocalStorage) {
+        localStorage.setItem(`${options.localStorageKey}-panels`, JSON.stringify(panels))
+    }
+}
 
 export { onAddPanel, findPanelFunc, findPanel };
